Move SNMP config commands to separated file
